% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{barplots_}
\alias{barplots_}
\title{barplots}
\usage{
barplots_(plateform = "IlluminaHumanMethylation450k", tab = NULL, man = NULL)
}
\arguments{
\item{plateform}{plateforme name}

\item{tab}{cpg table}

\item{man}{manifest}
}
\value{
plot
}
\description{
Display repartition of DMPs/DMRs annotation
}
