% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{barplots_450k}
\alias{barplots_450k}
\title{barplots_450k}
\usage{
barplots_450k(tab = NULL, man = NULL)
}
\arguments{
\item{tab}{cpg table}

\item{man}{manifest}
}
\value{
plot
}
\description{
Display repartition of DMPs/DMRs annotation
}
