% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{barplots_dist_to_tss}
\alias{barplots_dist_to_tss}
\title{barplots_dist_to_tss}
\usage{
barplots_dist_to_tss(dmps, bin = 50)
}
\arguments{
\item{dmps}{data.frame of dmps}

\item{bin}{size of bin}
}
\value{
plot
}
\description{
Display barplot of deltabetas distribution around TSS
}
