% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{barplots_EPIC}
\alias{barplots_EPIC}
\title{barplots_EPIC}
\usage{
barplots_EPIC(tab = NULL, man = NULL)
}
\arguments{
\item{tab}{cpg table}

\item{man}{manifest}
}
\value{
plot
}
\description{
Display repartition of DMPs/DMRs annotation
}
