% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\docType{class}
\name{Betas-class}
\alias{Betas-class}
\title{Class Betas}
\description{
This class extend SummarizedExperiment, with new methods, it made it specific for betas values.
}
\seealso{
\code{\link{SummarizedExperiment-class}} : The parent class

Other data container: 
\code{\link{Mvals-class}}
}
\concept{data container}
