% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{colData,Betas-method}
\alias{colData,Betas-method}
\title{Method for extracting column data from a "Betas" object}
\usage{
\S4method{colData}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object.}
}
\value{
A data frame containing the extracted column data.
}
\description{
This method extracts column data from a "Betas" object and returns it as a data frame.
}
