% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{colData,Mvals-method}
\alias{colData,Mvals-method}
\title{Method for extracting column data from a "Mvals" object}
\usage{
\S4method{colData}{Mvals}(x)
}
\arguments{
\item{x}{A "Mvals" object.}
}
\value{
A data frame containing the extracted column data.
}
\description{
This method extracts column data from a "Mvals" object and returns it as a data frame.
}
