% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{colData,NULL-method}
\alias{colData,NULL-method}
\title{Method for extracting column data from a "NULL" object}
\usage{
\S4method{colData}{`NULL`}(x)
}
\arguments{
\item{x}{A "NULL" object.}
}
\value{
NULL, as there is no column data to extract from a "NULL" object.
}
\description{
This method is used to extract column data from a "NULL" object. In this case, it always returns NULL, indicating that there is no column data to extract from a NULL object.
}
