% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{colData}
\alias{colData}
\title{Generic function to extract column data from an object}
\usage{
colData(x, ...)
}
\arguments{
\item{x}{An object from which to extract column data.}
}
\value{
The extracted column data.
}
\description{
Generic function to extract column data from an object
}
