% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cpg_excl}
\alias{cpg_excl}
\title{Filter probes from list of probes}
\usage{
cpg_excl(
  filters = NULL,
  plateform = "IlluminaHumanMethylationEPIC",
  snp = TRUE,
  cross = TRUE,
  xy = TRUE
)
}
\arguments{
\item{filters}{file containing probe list}

\item{plateform}{plateform}
}
\value{
vector
}
\description{
Filter probes from list of probes
}
