% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.r
\name{CpGNAexcl}
\alias{CpGNAexcl}
\title{CpG with too many NA}
\usage{
CpGNAexcl(betas, nalimit = 0.2)
}
\arguments{
\item{betas}{matrix of betas}

\item{nalimit}{maximum proportion of NA accepted}
}
\value{
List of probes to exclude
}
\description{
select probes with percentage of missing values superior to CpGlimit
}
