% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_steps.R
\name{create_report_index}
\alias{create_report_index}
\title{Create Report Home Page}
\usage{
create_report_index(report)
}
\arguments{
\item{report}{A \code{MethylkeyReport} object.}
}
\value{
The updated \code{MethylkeyReport} object.
}
\description{
Generate the report's \code{index.qmd} home page with analysis metadata and
session information.
}
\keyword{internal}
