% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs_plots.r
\name{density_plot}
\alias{density_plot}
\title{densityPlot of betas values}
\usage{
density_plot(
  meth,
  group = "sentrix_id",
  main = "",
  subset = 10000,
  xlab = "Beta values distribution",
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  ...
)
}
\arguments{
\item{main}{Plot title.}

\item{xlab}{Label for the x-axis.}

\item{pal}{Color palette vector.}

\item{dat}{A matrix of betas.}

\item{sampGroups}{A vector of group labels for each column in \code{dat}.}

\item{xlim}{x-axis limits.}

\item{ylim}{y-axis limits.}

\item{legend}{Whether to draw a legend.}
}
\value{
A ggplot2 object representing density estimates for each sample.
}
\description{
This function generates a ggplot2 density plot for matrix columns.
Each column is represented by a density line and colored by
\code{sampGroups}.
}
