% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{.dmrs.unify}
\alias{.dmrs.unify}
\title{.dmrs.unify Function}
\usage{
.dmrs.unify(x, dt, t)
}
\arguments{
\item{x}{An object.}

\item{dt}{A data frame containing DMR information.}

\item{t}{A character indicating the DMR tool.}
}
\value{
A data frame with unified DMR information.
}
\description{
\code{.dmrs.unify} Function
}
\details{
Description of the function \code{.dmrs.unify}.
}
