% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{dot_plot}
\alias{dot_plot}
\title{Create a Dot Plot for Gene-Term Associations}
\usage{
dot_plot(df)
}
\arguments{
\item{df}{A data frame containing gene-term associations with the following
required columns:
\itemize{
\item \code{P.value}: Numeric, p-values for the terms.
\item \code{gene_count}: Numeric, count of genes associated with each term.
\item \code{Term}: Character, names of the terms.
\item \code{Odds.Ratio}: Numeric, odds ratio values.
}}
}
\value{
A ggplot object representing the dot plot.
}
\description{
This function generates a dot plot that visualizes gene-term associations
based on odds ratio, gene count, and statistical significance.
}
