% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{estimate_PCA_corr}
\alias{estimate_PCA_corr}
\title{Estimate PCA Correlation}
\usage{
estimate_PCA_corr(pca, pdata, nPC)
}
\arguments{
\item{pca}{An object of class 'prcomp' containing PCA results.}

\item{pdata}{A data frame containing variables for which the correlation with PCs will be estimated.}

\item{nPC}{Number of principal components to include in the analysis.}
}
\value{
A matrix of p-values representing the correlation significance for each variable with each PC.
}
\description{
Computes correlation statistics for the contribution of variables to principal components (PC) in a PCA analysis.
}
