% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSheet.r
\name{formatSampleSheet}
\alias{formatSampleSheet}
\title{Format and rename columns in a sample sheet data frame}
\usage{
formatSampleSheet(sampleSheet)
}
\arguments{
\item{sampleSheet}{A data frame representing the sample sheet.}
}
\value{
A processed and formatted sample sheet data frame.
}
\description{
This function formats and renames columns in a sample sheet data frame to ensure consistency and compatibility with downstream analysis. It performs the following operations:
\itemize{
\item Converts column names to lowercase.
\item Renames the "samples" or "sample_id" column to "samples".
\item Renames the "barcode" or "basename" column to "barcode".
\item Ensures that the "barcode" and "samples" columns exist in the sample sheet.
\item Converts column names to valid R variable names.
\item Converts character columns to factors.
\item Separates the "barcode" column into "sentrix_id" and "sentrix_position" columns.
\item Arranges the rows by the "barcode" column.
}
}
