% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{get_annotated_manifest}
\alias{get_annotated_manifest}
\title{get_annotated_manifest}
\usage{
get_annotated_manifest(se, annot = NULL)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{annot}{annotation optionnelle}
}
\value{
DataFrame annoté
}
\description{
get_annotated_manifest
}
