% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{get_betas}
\alias{get_betas}
\title{Generic function to retrieve beta values from a "Betas" object}
\usage{
get_betas(x, masked = FALSE, na = FALSE, sex = TRUE)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve beta values.}

\item{masked}{Logical. If FALSE (default), filter out masked values.}

\item{na}{Logical. If FALSE (default), filter out NA values.}

\item{sex}{Logical. If TRUE (default), retain sex chromosomes data.}
}
\value{
A matrix of beta values, optionally filtered based on
the specified conditions.
}
\description{
This is a generic function to retrieve beta values from a "Betas" object.
It provides options to filter the data based on certain conditions.
}
