% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{get_betas_ranges,Betas-method}
\alias{get_betas_ranges,Betas-method}
\title{Method for getting beta value ranges from a Betas object}
\usage{
\S4method{get_betas_ranges}{Betas}(x, masked = FALSE, na = FALSE, sex = TRUE)
}
\arguments{
\item{x}{A Betas object from which to retrieve beta value ranges.}

\item{masked}{Logical indicating whether to include masked CpGs
(default is FALSE).}

\item{na}{Logical indicating whether to include CpGs marked as NA
(default is FALSE).}

\item{sex}{Logical indicating whether to include sex chromosomes
(default is TRUE).}
}
\value{
A GRanges object containing the filtered beta value ranges.
}
\description{
This method retrieves beta value ranges from a Betas object,
applying filters based on the provided parameters for masked CpGs,
NA values, and sex chromosomes. It returns a GRanges object
containing the filtered beta value ranges.
}
