% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_delta_betas}
\alias{get_delta_betas}
\title{Calculate Delta betas between two groups}
\usage{
get_delta_betas(betas, design, factor_prefix, level_col)
}
\arguments{
\item{betas}{array of betas values}

\item{design}{design matrix}

\item{factor_prefix}{prefix for factor columns}

\item{level_col}{column for the level of interest}
}
\value{
vector
}
\description{
Calculate Delta betas between two groups
}
