% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{get_genes}
\alias{get_genes}
\alias{get_genes,MethylResultSet-method}
\title{Extract Unique Genes from Significant DMPs and DMRs
Retrieves a list of unique gene symbols associated with significant DMPs and
DMRs for a given contrast, based on specified significance thresholds and DMR
tools.}
\usage{
get_genes(
  x,
  index,
  tools = c("dmrcate", "ipdmr", "combp", "dmrff"),
  q = 0.05,
  deltabetas = 3,
  min_cpgs = 2
)

\S4method{get_genes}{MethylResultSet}(
  x,
  index,
  tools = c("dmrcate", "ipdmr", "combp", "dmrff"),
  q = 0.05,
  deltabetas = 3,
  min_cpgs = 2
)
}
\arguments{
\item{x}{A \code{MethylResultSet} object containing DMP and DMR results.}

\item{index}{Integer index or character name of the contrast to analyze.}

\item{tools}{Character vector of DMR tools to consider for DMR-based gene
extraction. Default is \code{c("dmrcate", "ipdmr", "combp", "dmrff")}.}

\item{q}{Numeric value specifying the FDR threshold for significance
(default \code{0.05}).}

\item{deltabetas}{Numeric value specifying the minimum absolute
delta beta threshold for significance (default \code{0.3}).}

\item{min_cpgs}{Integer specifying the minimum number of CpGs
required in a DMR for it to be considered significant (default \code{2}).}
}
\value{
A character vector of unique gene symbols associated with
significant DMPs and DMRs based on the specified criteria.
}
\description{
Extract Unique Genes from Significant DMPs and DMRs
Retrieves a list of unique gene symbols associated with significant DMPs and
DMRs for a given contrast, based on specified significance thresholds and DMR
tools.
}
\details{
The function performs the following steps:
\itemize{
\item Extracts significant DMPs based on the specified FDR and
delta beta thresholds, and collects associated gene symbols.
\item Extracts significant DMRs based on the specified DMR tools,
FDR threshold, delta beta threshold, and minimum CpG count,
and collects associated gene symbols.
\item Combines gene symbols from both DMPs and DMRs, removes duplicates,
and returns a unique list of gene symbols.
}
}
\section{Functions}{
\itemize{
\item \code{get_genes(MethylResultSet)}: Method for MethylResultSet

}}
