% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{get_illumina_annotation}
\alias{get_illumina_annotation}
\title{get_illumina_annotation}
\usage{
get_illumina_annotation(plateform)
}
\arguments{
\item{plateform}{Type of plateform}
}
\value{
A data.frame annotation
}
\description{
get_illumina_annotation
}
