% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{get_manifest}
\alias{get_manifest}
\title{getManifest}
\usage{
get_manifest(plateform)
}
\arguments{
\item{plateform}{plateform}
}
\value{
annotated manifest
}
\description{
retrieve infinium annotations
}
