% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{get_mvals}
\alias{get_mvals}
\title{Generic function to retrieve M-values from an "Mvals" object}
\usage{
get_mvals(x, ...)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve M-values.}
}
\value{
A matrix of M-values.
}
\description{
This is a generic function to retrieve M-values from an "Mvals" object.
It provides a consistent interface for accessing M-values
across different classes.
}
