% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_create.R
\name{get_report_subdir}
\alias{get_report_subdir}
\title{Get Report Subdirectory}
\usage{
get_report_subdir(report)
}
\arguments{
\item{report}{A \code{MethylkeyReport} object.}
}
\value{
Character string with full path.
}
\description{
Returns the full path to the report's subdirectory.
}
\keyword{internal}
