% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{get_zhou_manifest}
\alias{get_zhou_manifest}
\title{getZhouManifest}
\usage{
get_zhou_manifest(plateform)
}
\arguments{
\item{plateform}{Type of plateform}
}
\value{
A data.frame manifest
}
\description{
getZhouManifest
}
