% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{getAnnotedManifest}
\alias{getAnnotedManifest}
\title{getAnnotedManifest}
\usage{
getAnnotedManifest(plateform)
}
\arguments{
\item{plateform}{plateform}
}
\value{
annotated manifest
}
\description{
retrieve infinium annotations
}
