% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getBetas,Betas-method}
\alias{getBetas,Betas-method}
\title{Method for retrieving beta values from a "Betas" object}
\usage{
\S4method{getBetas}{Betas}(x, masked = FALSE, na = FALSE, sex = TRUE)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve beta values.}

\item{masked}{Logical. If FALSE (default), filter out masked values.}

\item{na}{Logical. If FALSE (default), filter out NA values.}

\item{sex}{Logical. If TRUE (default), retain sex chromosomes data.}
}
\value{
A matrix of beta values, optionally filtered based on the specified conditions.
}
\description{
This method retrieves beta values from a "Betas" object, optionally filtering the data based on the specified conditions.
}
