% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getBetasRanges}
\alias{getBetasRanges}
\title{Create generic function for getting beta value ranges}
\usage{
getBetasRanges(x, masked = FALSE, na = FALSE, sex = TRUE)
}
\arguments{
\item{x}{A Betas object.}

\item{masked}{Logical indicating whether to include masked CpGs (default is FALSE).}

\item{na}{Logical indicating whether to include CpGs marked as NA (default is FALSE).}

\item{sex}{Logical indicating whether to include sex chromosomes (default is TRUE).}
}
\description{
Create a generic function for getting beta value ranges
}
\details{
This function sets up a generic function, \code{getBetasRanges}, for getting beta value ranges from a Betas object.
}
