% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getDefaultProbeList}
\alias{getDefaultProbeList}
\title{getDefaultProbeList}
\usage{
getDefaultProbeList(
  plateform = "IlluminaHumanMethylationEPIC",
  SNP = TRUE,
  CROSS = TRUE,
  XY = TRUE
)
}
\arguments{
\item{filters}{file containing probe list}
}
\value{
vector
}
\description{
get default list of probes to filter according to the plateform
}
