% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r, R/utils.r
\name{getDeltaBetas}
\alias{getDeltaBetas}
\title{getDeltaBetas}
\usage{
getDeltaBetas(betas, group, case = "TT", control = "NT")

getDeltaBetas(betas, group, case = "TT", control = "NT")
}
\arguments{
\item{betas}{array of betas values}

\item{group}{group}

\item{case}{case}

\item{control}{control}
}
\value{
vector

vector
}
\description{
calculate Delta betas

calculate Delta betas between two groups
}
