% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getDeltaBetas2}
\alias{getDeltaBetas2}
\title{Calculate delta betas between all groups}
\usage{
getDeltaBetas2(betas, group)
}
\arguments{
\item{betas}{A matrix of methylation beta values, where rows represent CpG sites and columns represent samples.}

\item{group}{A factor or character vector specifying group membership for each sample.}
}
\value{
A data frame containing delta betas for each pair of groups.
}
\description{
This function calculates delta betas (changes in methylation levels) between different groups based on methylation beta values. It computes the difference in methylation levels for all possible combinations of groups.
}
