% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{getDMPranges}
\alias{getDMPranges}
\title{Create generic function for retrieving DMP ranges}
\usage{
getDMPranges(x, group = 1, q = 0.5)
}
\arguments{
\item{x}{A MethylResultSet object.}

\item{group}{The group number for which to retrieve DMPs (default is 1).}

\item{q}{The threshold for adjusted p-value (default is 0.5).}
}
\description{
Create generic function for retrieving differentially methylated position (DMP) ranges
}
\details{
This function sets up a generic function, \code{getDMPranges}, for retrieving ranges of differentially methylated positions (DMPs) from a MethylResultSet object.
}
