% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{getDMPs}
\alias{getDMPs}
\title{Create generic function for retrieving DMPs}
\usage{
getDMPs(x, group = 1)
}
\arguments{
\item{x}{A MethylResultSet object.}

\item{group}{The group number for which to retrieve DMPs (default is 1).}
}
\description{
Create generic function for retrieving differentially methylated positions (DMPs)
}
\details{
This function sets up a generic function, \code{getDMPs}, for retrieving differentially methylated positions (DMPs) from a MethylResultSet object.
}
