% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{getDMRs}
\alias{getDMRs}
\alias{getDMRs,MethylResultSet-method}
\title{getDMRs Generic Function}
\usage{
getDMRs(x)

\S4method{getDMRs}{MethylResultSet}(x)
}
\arguments{
\item{x}{An object of class \code{MethylResultSet}.}
}
\value{
A result.

A data frame containing unified DMRs.
}
\description{
\code{getDMRs} Generic Function

Description of the method for extracting DMRs from a MethylResultSet object.
}
\details{
Description of the generic function \code{getDMRs}.
}
