% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{getInfiniumAnnotation}
\alias{getInfiniumAnnotation}
\title{getInfiniumAnnotation}
\usage{
getInfiniumAnnotation(plateform, genome = "hg19")
}
\arguments{
\item{plateform}{plateform}

\item{genome}{genome}
}
\value{
data.frame
}
\description{
retrieve infinium annotations
}
