% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getManifest}
\alias{getManifest}
\title{getAnnotedManifest}
\usage{
getManifest(plateform)
}
\arguments{
\item{plateform}{plateform}
}
\value{
annotated manifest
}
\description{
retrieve infinium annotations
}
