% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getMask}
\alias{getMask}
\title{Retrieve a masking matrix based on the Illumina methylation platform}
\usage{
getMask(mat)
}
\arguments{
\item{mat}{A data matrix representing methylation data.}
}
\value{
A masking matrix with relevant information for filtering CpG probes.
}
\description{
This function retrieves a masking matrix based on the Illumina methylation platform specified. The masking matrix is used for filtering and annotation of CpG probes.
}
