% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getMvals,Betas-method}
\alias{getMvals,Betas-method}
\title{Compute M-values from a "Betas" object}
\usage{
\S4method{getMvals}{Betas}(
  x,
  grp = "grp",
  sva = NULL,
  win = TRUE,
  sex = FALSE,
  sentrix2remove = list()
)
}
\arguments{
\item{x}{A "Betas" object from which to compute M-values.}

\item{grp}{A character vector specifying the group variable for differential analysis.}

\item{sva}{A character vector specifying the surrogate variable analysis (SVA) variable to include in the model. If NULL, SVA is skipped.}

\item{win}{A logical indicating whether Winsorization should be applied to the beta values. Default is TRUE.}

\item{sex}{A logical indicating whether sex chromosome data should be included. Default is FALSE.}

\item{sentrix2remove}{A list of Sentrix IDs to remove from the analysis. Default is an empty list.}
}
\value{
A "Mvals" object containing computed M-values.
}
\description{
This method computes M-values from a "Betas" object, which represents beta values.
}
