% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getMvals,Mvals-method}
\alias{getMvals,Mvals-method}
\title{Method for retrieving M-values from an "Mvals" object}
\usage{
\S4method{getMvals}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve M-values.}
}
\value{
A matrix of M-values.
}
\description{
This method retrieves M-values from an "Mvals" object.
}
