% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{getMvals}
\alias{getMvals}
\title{Generic function to retrieve M-values from an "Mvals" object}
\usage{
getMvals(x, ...)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve M-values.}
}
\value{
A matrix of M-values.
}
\description{
This is a generic function to retrieve M-values from an "Mvals" object.
}
