% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getPlateform}
\alias{getPlateform}
\title{getPlateform}
\usage{
getPlateform(betas)
}
\arguments{
\item{betas}{betas array}
}
\value{
plateform
}
\description{
guess illumina plateform based on probes names and probes number
}
