% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{getResults}
\alias{getResults}
\title{Create generic function for retrieving methylation analysis results}
\usage{
getResults(x, index, tools = c("dmrcate", "ipdmr"), ...)
}
\arguments{
\item{x}{A MethylResultSet object.}

\item{index}{Index or indices specifying the samples for which to retrieve results.}

\item{tools}{Character vector specifying the analysis tools to use (default is c("dmrcate", "ipdmr")).}

\item{...}{Additional arguments to be passed to the analysis functions.}
}
\description{
Create generic function for retrieving methylation analysis results
}
\details{
This function sets up a generic function, \code{getResults}, for retrieving methylation analysis results from a MethylResultSet object.
}
