% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{gtf2GenomicRanges}
\alias{gtf2GenomicRanges}
\title{gtf2GenomicRanges}
\usage{
gtf2GenomicRanges(file)
}
\arguments{
\item{file}{gtf file}
}
\value{
genomic ranges
}
\description{
transform gtf file to genomic ranges
}
