% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{h_mean}
\alias{h_mean}
\title{h_mean
Calculate the harmonic mean of a numeric vector}
\usage{
h_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove NA values
before calculation (default: TRUE)}
}
\value{
The harmonic mean of the input vector
}
\description{
h_mean
Calculate the harmonic mean of a numeric vector
}
