% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.r
\name{impute_na}
\alias{impute_na}
\title{impute missing values}
\usage{
impute_na(betas, nalimit = 0.2)
}
\arguments{
\item{betas}{matrix of betas}

\item{nalimit}{maximum proportion of NA accepted}
}
\value{
betas matrix
}
\description{
impute missing values with pamr for probes with less than 'nalimit'
proportion of NA values
}
