% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.r
\name{imputeNA}
\alias{imputeNA}
\title{impute missing values}
\usage{
imputeNA(betas, nalimit)
}
\arguments{
\item{betas}{matrix of betas}

\item{nalimit}{maximum proportion of NA accepted}
}
\value{
betas matrix
}
\description{
impute missing values with parmr for probes with less than 'nalimit' proportion of NA values
}
