% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_steps.R
\name{load_template}
\alias{load_template}
\title{Template Rendering Utilities}
\usage{
load_template(template_name)
}
\arguments{
\item{template_name}{Character. Name of template file (with .template extension).}
}
\value{
Character string with template content.
}
\description{
Internal functions for loading and rendering .qmd templates with parameter substitution.
}
\keyword{File}
\keyword{Load}
\keyword{Read}
\keyword{Template}
\keyword{a}
\keyword{file}
\keyword{from}
\keyword{inst/templates/}
\keyword{internal}
\keyword{template}
