% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{m_leastsquare2}
\alias{m_leastsquare2}
\title{m_leastsquare2}
\usage{
m_leastsquare2(mval, pdata, model)
}
\arguments{
\item{mval}{mvalues array}

\item{pdata}{sampleSheet}

\item{model}{model}
}
\value{
regression
}
\description{
least square method, return the whole table
}
