% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{m_leastsquare_ruv}
\alias{m_leastsquare_ruv}
\title{m_leastsquare_ruv}
\usage{
m_leastsquare_ruv(mval, pdata, model)
}
\arguments{
\item{mval}{mvalues array}

\item{pdata}{sampleSheet}

\item{model}{model}
}
\value{
regression
}
\description{
least square method with ruv correction, return the whole table
}
