% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{m_logistic}
\alias{m_logistic}
\title{m_logistic}
\usage{
m_logistic(mval, pdata, variables, fdr, ncore = 4)
}
\arguments{
\item{mval}{mvalues array}

\item{pdata}{sampleSheet}

\item{fdr}{fdr}

\item{ncore}{number of core (4)}

\item{model}{variables (vector)}
}
\value{
regression
}
\description{
conditionnal logistic regression, return direcly a toptable
}
