% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{m_regression}
\alias{m_regression}
\title{m_regression}
\usage{
m_regression(mval, pdata, model, method = "ls", niter = 50, ncore = 4)
}
\arguments{
\item{mval}{mvalues array}

\item{pdata}{sampleSheet}

\item{model}{model}

\item{method}{ls,ls_ruv,robust,logistic}
}
\value{
regression
}
\description{
Display betas in a violin plot by group
}
