% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{m_robust}
\alias{m_robust}
\title{m_robust}
\usage{
m_robust(mval, pdata, model, niter = 50)
}
\arguments{
\item{mval}{mvalues array}

\item{pdata}{sampleSheet}

\item{model}{model}

\item{niter}{number of iteration (50)}
}
\value{
regression
}
\description{
least square rosbust method, return the whole table
}
