% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{makeFormula}
\alias{makeFormula}
\title{makeFormula}
\usage{
makeFormula(x, variables)
}
\arguments{
\item{x}{x}

\item{model}{variables (vector)}
}
\value{
regression
}
\description{
make formula for logistic regression
}
