% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{makepca}
\alias{makepca}
\title{PCA analysis}
\usage{
makepca(mvals, pdata)
}
\arguments{
\item{mvals}{matrix of m values}

\item{pdata}{sampleSheet of data}
}
\value{
betas matrix
}
\description{
Estimate correlation covariates vs. PCA axes
}
