% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makeSummarizedExperimentFromBetas}
\alias{makeSummarizedExperimentFromBetas}
\title{makeSummarizedExperimentFromBetas}
\usage{
makeSummarizedExperimentFromBetas(betas, sampleSheet)
}
\arguments{
\item{betas}{betas array}

\item{sampleSheet}{data.frame}
}
\value{
SummarizedExperiment
}
\description{
create a summarized experiment
}
