% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{manhattan}
\alias{manhattan}
\title{Manhattan}
\usage{
manhattan(df, sig = 5e-08)

manhattan(df, sig = 5e-08)
}
\arguments{
\item{df}{A data frame containing GWAS data with columns chr (chromosome), pos (position), and P.Value (p-value).}

\item{sig}{The significance threshold for highlighting points on the plot (default is 5e-8).}

\item{tab}{dmps table (dataframe)}
}
\value{
plot

A Manhattan plot visualizing GWAS data.
}
\description{
Manhattan plot of DMPs

This function generates a Manhattan plot for visualizing genome-wide association study (GWAS) data. It takes a data frame containing chromosome (chr), position (pos), and p-value (P.Value) columns as input.
}
\details{
The table must contain a column 'chr' for chromosome
and a column pos for the chromosomique position.
}
