% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylkey-package.r
\docType{package}
\name{methylkey-package}
\alias{methylkey-package}
\alias{methylkey}
\alias{MethylResultSet}
\alias{add_combp}
\alias{add_dmrcate}
\alias{add_dmrff}
\alias{add_ipdmr}
\alias{dmrtools_dmrcate}
\alias{dmrtools_dmrff}
\alias{dmrtools_ipdmr}
\alias{dmrtools_combp}
\alias{as_dataframe}
\alias{barplots_dist_to_tss}
\alias{best_probe_by_gene}
\alias{circosplot}
\alias{colData}
\alias{cpgislands_plot}
\alias{density_plot}
\alias{my_scatter_plot}
\alias{my_violin}
\alias{manhattan}
\alias{estimate_pca_corr}
\alias{format_sample_sheet}
\alias{get_betas}
\alias{get_betas_ranges}
\alias{get_dmps}
\alias{get_dmps_ranges}
\alias{get_dmrs}
\alias{get_dmrs_ranges}
\alias{get_genes}
\alias{get_manifest}
\alias{get_mask}
\alias{get_model_name}
\alias{get_mvals}
\alias{get_number_of_dmrs}
\alias{get_plateform}
\alias{get_results}
\alias{makepca}
\alias{minfi2betas}
\alias{sesame2betas}
\alias{nbprobes}
\alias{nbsamples}
\alias{nbsentrix}
\alias{new_betas}
\alias{plot_background}
\alias{plot_channel}
\alias{switch}
\alias{plot_channels}
\alias{plot_channels2}
\alias{plot_detection}
\alias{plot_na}
\alias{plot}
\alias{pca_contribution}
\alias{probes}
\alias{samples}
\alias{sankeydot_plot}
\alias{sentrix}
\alias{setColData}
\alias{subset_mval}
\alias{to_geo_submission}
\alias{tools_density}
\alias{tools_venn}
\alias{variables}
\alias{violin_plot}
\alias{volcano}
\alias{volcano_label_probes}
\title{methylkey: DNA Methylation Analysis from Illumina Arrays}
\description{
\code{methylkey} provides a comprehensive Bioconductor package for analysis
of DNA methylation
data from Illumina methylation arrays (27k, 450k, EPIC, and mouse arrays).
}
\details{
The package offers:
\itemize{
\item Preprocessing pipelines using \code{sesame} and \code{minfi} packages
\item Quality control and visualization functions
\item Differential methylation analysis (DMP and DMR detection)
\item Support for multiple array platforms and normalization methods
\item Integration with SummarizedExperiment for data management
}

Main workflow:
\enumerate{
\item Prepare a sample sheet with sample metadata and barcode information
\item Load IDAT files using \code{\link{sesame2Betas}}
or \code{\link{minfi2Betas}}
\item Perform quality control and visualization
\item Run differential methylation analysis with \code{\link{methyldiff}}
\item Generate reports with \code{\link{methylkey_report}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IARCbioinfo/methylkey}
  \item Report bugs at \url{https://github.com/IARCbioinfo/methylkey/issues}
}

}
\author{
\strong{Maintainer}: Vincent Cahais \email{cahaisv@iarc.fr} (\href{https://orcid.org/0000-0001-5530-4368}{ORCID})

Authors:
\itemize{
  \item Vincent Cahais \email{cahaisv@iarc.fr} (\href{https://orcid.org/0000-0001-5530-4368}{ORCID})
}

}
