% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.r
\name{minfi2Betas}
\alias{minfi2Betas}
\title{Convert Illumina IDAT Files to Beta Values Using minfi}
\usage{
minfi2Betas(
  idat = NULL,
  sampleSheet = NULL,
  na = 0.2,
  compositeCellType = "",
  pval = 0.2
)
}
\arguments{
\item{idat}{The path to the directory containing Illumina IDAT files.}

\item{sampleSheet}{A data frame containing sample information, including Basename, sex, and age.}

\item{na}{The maximum proportion of missing values allowed in beta values. Must be between 0 and 1. Default is 0.2.}

\item{compositeCellType}{A composite cell type to estimate cell counts, e.g., "Blood," "CordBloodCombined," etc.}
}
\value{
An object of class "MethyLumiMethyLumiSet" containing beta values and associated metadata, including quality control statistics and plots.
}
\description{
This function takes Illumina IDAT files, processes them using the minfi package, and returns beta values (DNA methylation values) along with quality control statistics and plots.
}
