% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{mk_annotatr}
\alias{mk_annotatr}
\title{make annotation with annotatr}
\usage{
mk_annotatr(regions, annot = NULL, genome = "hg19")
}
\arguments{
\item{regions}{GRanges}

\item{annot}{annotation database}

\item{genome}{genome}
}
\value{
annotation file
}
\description{
make annotation with annotatr
}
