% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{my_dmrplot}
\alias{my_dmrplot}
\title{Create a DMR plot for visualizing beta values with confidence intervals}
\usage{
my_dmrplot(dt, group, title)
}
\arguments{
\item{dt}{A data frame containing the required data.}

\item{group}{A grouping variable for distinguishing different groups in the plot.}

\item{title}{A title for the plot.}
}
\value{
A ggplot2 plot displaying beta values with confidence intervals for different groups.
}
\description{
This function generates a DMR (Differentially Methylated Region) plot for visualizing beta values with confidence intervals. It takes a data frame 'dt' containing the required data, a grouping variable 'group', and a title for the plot.
}
