% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{my_venn_}
\alias{my_venn_}
\title{Create a Venn diagram to visualize intersections of probe IDs}
\usage{
my_venn_(dt, a)
}
\arguments{
\item{dt}{A data frame containing the data.}

\item{a}{A list of set names to be intersected.}
}
\value{
A ggvenn Venn diagram showing the intersections between sets.
}
\description{
This function generates a Venn diagram to visualize the intersections of probe IDs from different sets. It takes a data frame 'dt' as input and a list 'a' containing the names of sets to be intersected.
}
