% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{my_venn_oriented}
\alias{my_venn_oriented}
\title{Create a Venn diagram to visualize intersections of IDs for different tools}
\usage{
my_venn_oriented(dt)
}
\arguments{
\item{dt}{A data frame containing the data.}
}
\value{
A ggvenn Venn diagram showing the intersections between IDs obtained from different tools.
}
\description{
This function generates a Venn diagram to visualize the intersections of IDs obtained from different analysis tools. It takes a data frame 'dt' as input and automatically extracts IDs for each tool before passing them to the 'my_venn_' function.
}
