% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{my_volcanohmean}
\alias{my_volcanohmean}
\title{Create a volcano plot to visualize differential methylation analysis results with harmonized mean}
\usage{
my_volcanohmean(dt)
}
\arguments{
\item{dt}{A data frame containing the data for differential methylation analysis.}
}
\value{
A ggplot2 volcano plot showing differential methylation analysis results with modified mean calculation.
}
\description{
This function generates a volcano plot to visualize differential methylation analysis results using ggplot2. It takes a data frame 'dt' as input, groups the data by 'dmrtool', 'ID', and 'fdr', calculates summary statistics with a modified 'mean.deltabetas' calculation, and creates a plot.
}
