% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbprobes,Betas-method}
\alias{nbprobes,Betas-method}
\title{Method for retrieving the number of probes from a "Betas" object}
\usage{
\S4method{nbprobes}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve the number of probes.}
}
\value{
The number of probes in the "Betas" object as an integer.
}
\description{
This method retrieves the number of probes from a "Betas" object.
}
