% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbprobes,Mvals-method}
\alias{nbprobes,Mvals-method}
\title{Method for retrieving the number of probes from an "Mvals" object}
\usage{
\S4method{nbprobes}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve the number of probes.}
}
\value{
The number of probes in the "Mvals" object as an integer.
}
\description{
This method retrieves the number of probes from an "Mvals" object.
}
