% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbprobes}
\alias{nbprobes}
\title{Generic function to retrieve the number of probes from an object}
\usage{
nbprobes(x)
}
\arguments{
\item{x}{An object from which to retrieve the number of probes.}
}
\value{
The number of probes as an integer.
}
\description{
This is a generic function to retrieve the number of probes from an object.
}
