% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsamples,Betas-method}
\alias{nbsamples,Betas-method}
\title{Method for retrieving the number of samples from a "Betas" object}
\usage{
\S4method{nbsamples}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve the number of samples.}
}
\value{
The number of samples in the "Betas" object.
}
\description{
This method retrieves the number of samples from a "Betas" object.
}
