% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsamples,Mvals-method}
\alias{nbsamples,Mvals-method}
\title{Method for retrieving the number of samples from an "Mvals" object}
\usage{
\S4method{nbsamples}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve the number of samples.}
}
\value{
The number of samples in the "Mvals" object.
}
\description{
This method retrieves the number of samples from an "Mvals" object.
}
