% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsentrix,Betas-method}
\alias{nbsentrix,Betas-method}
\title{Method for retrieving the number of Sentrix IDs from a "Betas" object}
\usage{
\S4method{nbsentrix}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve the number of Sentrix IDs.}
}
\value{
The number of Sentrix IDs as an integer.
}
\description{
This method retrieves the number of Sentrix IDs from a "Betas" object.
}
