% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsentrix,Mvals-method}
\alias{nbsentrix,Mvals-method}
\title{Method for retrieving the number of Sentrix IDs from an "Mvals" object}
\usage{
\S4method{nbsentrix}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve the number of Sentrix IDs.}
}
\value{
The number of Sentrix IDs as an integer.
}
\description{
This method retrieves the number of Sentrix IDs from an "Mvals" object.
}
