% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsentrix}
\alias{nbsentrix}
\title{Generic function to retrieve the number of Sentrix IDs from an object}
\usage{
nbsentrix(x)
}
\arguments{
\item{x}{An object from which to retrieve the number of Sentrix IDs.}
}
\value{
The number of Sentrix IDs as an integer.
}
\description{
This is a generic function to retrieve the number of Sentrix IDs from an object.
}
