% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{newBetas}
\alias{newBetas}
\title{Create a new "Betas" object from beta values and a sample sheet}
\usage{
newBetas(betas, sampleSheet, na)
}
\arguments{
\item{betas}{A matrix of beta values.}

\item{sampleSheet}{A data frame containing sample information.}

\item{na}{A numeric value, between 0 and 1, a probe is MASKED if its NA proportion is superior to na}
}
\value{
A "Betas" object containing the specified beta values, sample information, and MASK information as rowData.
}
\description{
This function creates a new "Betas" object from beta values and a sample sheet.
}
