% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_background}
\alias{plot_background}
\title{Create a scatter plot with custom annotations}
\usage{
plot_background(df)
}
\arguments{
\item{df}{A data frame containing the data to be plotted, with columns 'mean_oob_grn', 'mean_oob_red', and 'name'.}
}
\value{
A ggplot2 scatter plot with custom annotations.
}
\description{
This function generates a scatter plot with custom annotations using ggplot2. The plot displays data from a data frame, where 'mean_oob_grn' and 'mean_oob_red' are used as x and y coordinates, and 'name' is used for annotations.
}
