% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_channel_switch}
\alias{plot_channel_switch}
\title{Create a scatter plot to visualize channel switching}
\usage{
plot_channel_switch(df)
}
\arguments{
\item{df}{A data frame containing the data to be plotted, with columns 'InfI_switch_G2R' and 'InfI_switch_R2G'.}
}
\value{
A ggplot2 scatter plot for visualizing channel switching.
}
\description{
This function generates a scatter plot to visualize channel switching using ggplot2. It plots data from a data frame, where 'InfI_switch_G2R' and 'InfI_switch_R2G' are used as x and y coordinates.
}
