% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs_plots.r
\name{plot_channels,list-method}
\alias{plot_channels,list-method}
\title{Plot channels from a list}
\usage{
\S4method{plot_channels}{list}(x, num_positions = 1000)
}
\arguments{
\item{x}{A list where each element represents a set of channels.}
}
\value{
A ggplot2 boxplot representing the channels.
}
\description{
This method of \code{plot_channels} is specifically for plotting channels
when the input data is a list of channel sets. It extracts and organizes
the channels, creates a boxplot, and facets the plot by sentrix.
}
