% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs_plots.r
\name{plot_channels,RGChannelSet-method}
\alias{plot_channels,RGChannelSet-method}
\title{Plot channels from an RGChannelSet object}
\usage{
\S4method{plot_channels}{RGChannelSet}(x)
}
\arguments{
\item{x}{An RGChannelSet object containing microarray channel data.}
}
\value{
A ggplot2 plot with separate boxplots for red and green channels.
}
\description{
This method of \code{plot_channels} is specifically for plotting channels
when the input data is an RGChannelSet object. It creates boxplots
for red and green channels.
}
