% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{probes,Betas-method}
\alias{probes,Betas-method}
\title{Method for retrieving probe names from a "Betas" object}
\usage{
\S4method{probes}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve probe names.}
}
\value{
A character vector containing probe names.
}
\description{
This method retrieves probe names from a "Betas" object.
}
