% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{probes,Mvals-method}
\alias{probes,Mvals-method}
\title{Method for retrieving probe names from an "Mvals" object}
\usage{
\S4method{probes}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve probe names.}
}
\value{
A character vector containing probe names.
}
\description{
This method retrieves probe names from an "Mvals" object.
}
