% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{probes}
\alias{probes}
\title{Generic function to retrieve probe names from an object}
\usage{
probes(x)
}
\arguments{
\item{x}{An object from which to retrieve probe names.}
}
\value{
A character vector containing probe names.
}
\description{
This is a generic function to retrieve probe names from an object.
}
