% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSheet.r
\name{readSampleSheet}
\alias{readSampleSheet}
\title{readSampleSheet}
\usage{
readSampleSheet(
  path,
  samples = NULL,
  barcode = NULL,
  groups = NULL,
  sep = "\\t"
)
}
\arguments{
\item{path}{file location}

\item{barcode}{barcode column name}

\item{groups}{groups columns names (vector)}

\item{sep}{field separator ('\\t')}

\item{sample}{sample column name}
}
\value{
dataframe
}
\description{
Display betas in a violin plot by group
}
