% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.r
\name{replace_by_mean}
\alias{replace_by_mean}
\title{replace_by_mean}
\usage{
replace_by_mean(betas, groups)
}
\arguments{
\item{betas}{matrix of betas}

\item{group}{group name}
}
\value{
betas matrix
}
\description{
Replace missing values by mean for each group
}
