% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.r
\name{replaceByMean}
\alias{replaceByMean}
\title{replaceByMean}
\usage{
replaceByMean(betas, groups)
}
\arguments{
\item{betas}{matrix of betas}

\item{group}{group name}
}
\value{
betas matrix
}
\description{
Replace missing values by mean for each group
}
