% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{resultsToLong}
\alias{resultsToLong}
\title{Convert results to long format}
\usage{
resultsToLong(df)
}
\arguments{
\item{df}{A data frame containing the results.}
}
\value{
A data frame in long format.
}
\description{
Convert results to long format
}
\details{
This function takes a data frame containing results from different tools
(e.g., dmrcate, ipdmr, combp, dmrff) and converts it to long format.
}
