% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{samples,Betas-method}
\alias{samples,Betas-method}
\title{Method for retrieving sample information from a "Betas" object}
\usage{
\S4method{samples}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve sample information.}
}
\value{
Sample information as a data frame.
}
\description{
This method retrieves sample information from a "Betas" object.
}
