% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{samples,Mvals-method}
\alias{samples,Mvals-method}
\title{Method for retrieving sample information from an "Mvals" object}
\usage{
\S4method{samples}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve sample information.}
}
\value{
Sample information as a data frame.
}
\description{
This method retrieves sample information from an "Mvals" object.
}
