% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{samples}
\alias{samples}
\title{Generic function to retrieve sample information from an object}
\usage{
samples(x)
}
\arguments{
\item{x}{An object from which to retrieve sample information.}
}
\value{
Sample information, typically as a data frame or other suitable format.
}
\description{
This is a generic function to retrieve sample information from an object.
}
