% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{sankey_plot_}
\alias{sankey_plot_}
\title{Create a Sankey Plot with Colored Nodes}
\usage{
sankey_plot_(
  df,
  v_space = "auto",
  term_size = 3,
  gene_size = 1.5,
  palette = NULL
)
}
\arguments{
\item{df}{A data frame containing gene-term associations with the
following required columns:
\itemize{
\item \code{P.value}: Numeric, p-values for the terms.
\item \code{gene_count}: Numeric, count of genes associated with each term.
\item \code{Term}: Character, names of the terms.
\item \code{Odds.Ratio}: Numeric, odds ratio values.
\item \code{status}: Character, indicating if a node is "up", "down",
or another category.
}}

\item{term_size}{: text size for terms}

\item{gene_size}{: text size for genes}

\item{palette}{A named vector of colors.
The colors should be named according to the
categories in the \code{status} column, with special colors for "up" and "down".}
}
\value{
A ggplot object representing the Sankey diagram.
}
\description{
This function generates a Sankey plot from a given dataframe,
visualizing relationships
between genes and terms. The node colors are determined
by the provided palette.
}
