% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.r
\name{sankeydot_plot}
\alias{sankeydot_plot}
\title{Create a Combined Sankey and Dot Plot}
\usage{
sankeydot_plot(
  df,
  v_space = "auto",
  term_size = 3,
  gene_size = 1.5,
  palette = NULL
)
}
\arguments{
\item{df}{A data frame containing gene-term associations with the following
required columns:
\itemize{
\item \code{P.value}: Numeric, p-values for the terms.
\item \code{gene_count}: Numeric, count of genes associated with each term.
\item \code{Term}: Character, names of the terms.
\item \code{Odds.Ratio}: Numeric, odds ratio values.
\item \code{status}: Character, indicating node status
(e.g., "up", "down", or other terms).
}}

\item{palette}{A named vector of colors for the terms and statuses.
If \code{NULL},
the function generates a default palette using \code{viridis::viridis()},
with fixed colors for \code{"up"} (\verb{#317AC1}) and \code{"down"} (\verb{#E1A624}).}
}
\value{
A ggplot object combining a Sankey plot and a dot plot.
}
\description{
This function generates a composite visualization that includes both a
Sankey plot and a dot plot, allowing for an intuitive representation of
gene-term associations.
}
