% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrtools.r
\name{searchDMR_dmrff}
\alias{searchDMR_dmrff}
\title{Search for Differentially Methylated Regions (DMRs) using DMRff}
\usage{
searchDMR_dmrff(dmps, betas, maxgap = 1000)
}
\arguments{
\item{dmps}{A data frame containing CpG site information, including chromosome, position, statistical coefficient, standard deviation, p-value, and probe IDs.}

\item{betas}{A data frame containing methylation data, typically beta values, for the CpG sites.}

\item{maxgap}{The maximum gap between neighboring CpGs to consider when defining DMRs. Default is 1000.}
}
\value{
A data frame with information about identified DMRs, including chromosome, start and end positions, associated CpG probe IDs, and DMR tool name.
}
\description{
This function searches for Differentially Methylated Regions (DMRs) in DNA methylation data using the DMRff package. It takes a set of CpG sites with associated statistical information and methylation data and performs DMR analysis.
}
