% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{sentrix,Betas-method}
\alias{sentrix,Betas-method}
\title{Method for retrieving Sentrix IDs from a "Betas" object}
\usage{
\S4method{sentrix}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve Sentrix IDs.}
}
\value{
A character vector containing Sentrix IDs.
}
\description{
This method retrieves Sentrix IDs from a "Betas" object.
}
