% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{sentrix,Mvals-method}
\alias{sentrix,Mvals-method}
\title{Method for retrieving Sentrix IDs from an "Mvals" object}
\usage{
\S4method{sentrix}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve Sentrix IDs.}
}
\value{
A character vector containing Sentrix IDs.
}
\description{
This method retrieves Sentrix IDs from an "Mvals" object.
}
