% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{sentrix}
\alias{sentrix}
\title{Generic function to retrieve Sentrix IDs from an object}
\usage{
sentrix(x)
}
\arguments{
\item{x}{An object from which to retrieve Sentrix IDs.}
}
\value{
A character vector containing Sentrix IDs.
}
\description{
This is a generic function to retrieve Sentrix IDs from an object.
}
