% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\docType{class}
\name{List_OR_NULL-class}
\alias{List_OR_NULL-class}
\title{Define a Class Union}
\description{
Defines a class union accepting either a list or NULL. Used for optional
slots in S4 classes.
}
\keyword{classes}
