% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{setColData,Betas-method}
\alias{setColData,Betas-method}
\title{Method for setting column data for a "Betas" object}
\usage{
\S4method{setColData}{Betas}(x, pdata)
}
\arguments{
\item{x}{A "Betas" object to which to set the column data.}

\item{pdata}{A data frame containing the new column data.}
}
\value{
The modified "Betas" object with the updated column data.
}
\description{
This method sets the column data for a "Betas" object using the provided data frame.
}
