% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{setColData}
\alias{setColData}
\title{Generic function to set column data for an object}
\usage{
setColData(x, pdata)
}
\arguments{
\item{x}{An object to which to set the column data.}

\item{pdata}{A data frame containing the new column data.}
}
\value{
The modified object with the updated column data.
}
\description{
This is a generic function to set column data for an object. It defines
the interface for setting column data and can be used with different classes.
}
