% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_create.R
\name{show,MethylkeyReport-method}
\alias{show,MethylkeyReport-method}
\title{Show Method for MethylkeyReport}
\usage{
\S4method{show}{MethylkeyReport}(object)
}
\arguments{
\item{object}{A \code{MethylkeyReport} object.}
}
\description{
Display summary information about a MethylkeyReport object.
}
