% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{subset_mval}
\alias{subset_mval}
\title{Return a subset of Mvals data}
\usage{
subset_mval(se, samples2keep, probes2keep)
}
\arguments{
\item{a}{vector of Barcode corresponding to the samples to keep.}
}
\value{
The new subset Mval object.
}
\description{
This function return a subset of Mvals data with associated sampleSheet
}
