% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{to_geo_submission,Betas,character-method}
\alias{to_geo_submission,Betas,character-method}
\title{generate files to submit data to Geo database}
\usage{
\S4method{to_geo_submission}{Betas,character}(x, outfile = "betas2Geo.tsv")
}
\arguments{
\item{outfile}{file output name

importFrom assertthat asserthat
importFrom readr write_tsv}

\item{Betas}{a Betas object}
}
\description{
generate files to submit data to Geo database
}
