% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{to_geo_submission}
\alias{to_geo_submission}
\title{Generic function to extract column data from an object}
\usage{
to_geo_submission(x, outfile)
}
\arguments{
\item{x}{An object RGset or sdfs list.}

\item{outfile}{file output name}
}
\description{
Generic function to extract column data from an object
}
