% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{toGeoSubmission,Betas,character-method}
\alias{toGeoSubmission,Betas,character-method}
\title{generate files to submit data to Geo database}
\usage{
\S4method{toGeoSubmission}{Betas,character}(x, outfile = "betas2Geo.tsv")
}
\arguments{
\item{outfile}{file output name}

\item{Betas}{a Betas object}
}
\description{
generate files to submit data to Geo database
}
