% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{toGeoSubmission,RGChannelSet,character-method}
\alias{toGeoSubmission,RGChannelSet,character-method}
\title{generate files to submit data to Geo database}
\usage{
\S4method{toGeoSubmission}{RGChannelSet,character}(x, outfile = "rawdata2Geo.tsv")
}
\arguments{
\item{outfile}{file output name}

\item{RGset}{an RGset object}
}
\description{
generate files to submit data to Geo database
}
