% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_comparison_plot.r
\name{tools_density}
\alias{tools_density}
\title{Create a density plot of DMR lengths grouped by dmrtool}
\usage{
tools_density(dt)
}
\arguments{
\item{dt}{A data frame containing the data.}
}
\value{
A ggplot2 density plot showing DMR length distributions
grouped by 'dmrtool'.
}
\description{
This function generates a density plot of DMR
(Differentially Methylated Region) lengths grouped by 'dmrtool'.
It takes a data frame 'dt' as input, calculates the length of each DMR,
and then creates the density plot.
}
