% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{variables,Betas-method}
\alias{variables,Betas-method}
\title{Method for retrieving variable names from a "Betas" object}
\usage{
\S4method{variables}{Betas}(x)
}
\arguments{
\item{x}{A "Betas" object from which to retrieve variable names.}
}
\value{
A character vector containing variable names.
}
\description{
This method retrieves variable names (column names) from a "Betas" object.
}
