% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{variables,Mvals-method}
\alias{variables,Mvals-method}
\title{Method for retrieving variable names from an "Mvals" object}
\usage{
\S4method{variables}{Mvals}(x)
}
\arguments{
\item{x}{An "Mvals" object from which to retrieve variable names.}
}
\value{
A character vector containing variable names.
}
\description{
This method retrieves variable names (column names) from an "Mvals" object.
}
