% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{variables}
\alias{variables}
\title{Generic function to retrieve variable names from an object}
\usage{
variables(x)
}
\arguments{
\item{x}{An object from which to retrieve variable names.}
}
\value{
A character vector containing variable names.
}
\description{
This is a generic function to retrieve variable names (column names) from an object.
}
