% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{violin_plot1}
\alias{violin_plot1}
\title{violin_plot1}
\usage{
violin_plot1(betas, group, numPositions = 1000)
}
\arguments{
\item{betas}{matrix of betas}

\item{group}{variable group}

\item{numPositions}{number of random CpG used}
}
\value{
plot
}
\description{
Display betas in a violin plot by group
}
