% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{violin_plot3}
\alias{violin_plot3}
\title{violin_plot3}
\usage{
violin_plot3(betas, annot, group, feature, numPositions = 1000)
}
\arguments{
\item{betas}{matrix of betas}

\item{annot}{annotation table}

\item{group}{variable group}

\item{feature}{annotation feature to use.}

\item{numPositions}{number of random CpG used}
}
\value{
plot
}
\description{
Display betas in a violin plot by group
}
