% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_steps.R
\name{write_qmd_file}
\alias{write_qmd_file}
\title{Write QMD File}
\usage{
write_qmd_file(report, filename, content)
}
\arguments{
\item{report}{A \code{MethylkeyReport} object.}

\item{filename}{Character. Name of output .qmd file.}

\item{content}{Character. Rendered template content.}
}
\value{
Character path to written file (invisibly).
}
\description{
Write rendered .qmd content to report directory.
}
\keyword{internal}
