% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{add_dmrcate}
\alias{add_dmrcate}
\title{Create generic function for converting results to long format}
\usage{
add_dmrcate(x, fdr = 0.05, pcutoff = 0.02, maxgap = 1000, genome = "hg38")
}
\arguments{
\item{df}{A data frame containing results.}
}
\description{
Create generic function for converting results to long format
}
\details{
This function sets up a generic function, \code{resultsToLong}, for converting results to long format.
}
