% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\name{add_ipdmr}
\alias{add_ipdmr}
\title{Create generic function for adding IPDMR results}
\usage{
add_ipdmr(x, maxgap = 1000)
}
\arguments{
\item{x}{A data frame.}

\item{maxgap}{Maximum gap size for IPDMR regions (default is 1000).}
}
\description{
Create generic function for adding IPDMR results to a data frame
}
\details{
This function sets up a generic function, \code{add_ipdmr}, for adding IPDMR results to a data frame.
}
