% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.r
\name{bc_sva}
\alias{bc_sva}
\title{Batch Correction with SVA}
\usage{
bc_sva(mval, pdata, model)
}
\arguments{
\item{mval}{matrix of mvalues}

\item{pdata}{sample Sheet (dataframe)}

\item{model}{model to apply with sva (string) eg: "~group+gender"}
}
\value{
A matrix of batch corrected mvalues
}
\description{
This function will do batch correction on mvalues,
}
