% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{circosplot}
\alias{circosplot}
\title{circosplot}
\usage{
circosplot(ranges, genome)
}
\arguments{
\item{ranges}{GRanges object}

\item{genome}{(hg38,hg19,mm10)}
}
\value{
plot
}
\description{
Display betas in a violin plot by group
}
