% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{CpGexcl}
\alias{CpGexcl}
\title{CpGexcl}
\usage{
CpGexcl(
  filters = NULL,
  plateform = "IlluminaHumanMethylationEPIC",
  SNP = TRUE,
  CROSS = TRUE,
  XY = TRUE
)
}
\arguments{
\item{filters}{file containing probe list}

\item{plateform}{plateform}
}
\value{
vector
}
\description{
Filter probes from list of probes
}
