% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_delta_betas}
\alias{get_delta_betas}
\title{Calculate Delta betas between two groups}
\usage{
get_delta_betas(betas, group)
}
\arguments{
\item{betas}{array of betas values}

\item{group}{group}

\item{case}{case}

\item{control}{control}
}
\value{
vector
}
\description{
Calculate Delta betas between two groups
}
