% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{makepca}
\alias{makepca}
\title{PCA analysis}
\usage{
makepca(betas, pdata, variables = colnames(pdata), nPC = ncol(betas))
}
\arguments{
\item{betas}{matrix of betas}

\item{pdata}{sampleSheet of data}

\item{variables}{vector of variables names to include in the PCA}

\item{nPC}{maximum number of principal component to display}
}
\value{
betas matrix
}
\description{
Estimate correlation covariates vs. PCA axes
}
