% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{manhattan}
\alias{manhattan}
\title{Manhattan}
\usage{
manhattan(df, sig = 5e-08)
}
\arguments{
\item{sig}{cutoff significance}

\item{tab}{dmps table (dataframe)}
}
\value{
plot
}
\description{
Manhattan plot of DMPs
}
\details{
The table must contain a column 'chr' for chromosome
and a column pos for the chromosomique position.
}
