% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylkey.diff.r
\name{methyldiff}
\alias{methyldiff}
\title{Methylkey differential analysis}
\usage{
methyldiff(
  model = NULL,
  case = NULL,
  control = NULL,
  betas = NULL,
  pdata = NULL,
  sva = TRUE,
  method = "ls",
  niter = 50,
  ncore = 2,
  qval = 0.05,
  pcutoff = 0.2,
  pca_group = NULL,
  output = "methylkey",
  manifest = NULL,
  display = 500,
  plateform = NULL,
  genome = "hg19"
)
}
\arguments{
\item{model}{model to use for the analysis (string) eg: "~group+gender"}

\item{case}{name of case level}

\item{control}{name of control level}

\item{betas}{matrix of beta values}

\item{pdata}{SampleSheet (dataframe)}

\item{sva}{if TRUE, sva batch correction will be applied to mvalues}

\item{method}{method to use in regression (ls or robust)}

\item{niter}{number of iteration with robust method}

\item{ncore}{number of core use for regression analysis}

\item{qval}{qvalue cutoff for selecting significant dmps}

\item{pcutoff}{qvalue cutoff for selecting dmps used for dmr analysis}

\item{pca_group}{vector of variables tested in the PCA}

\item{output}{output folder}

\item{manifest}{annotation for th array}

\item{display}{number of dmps displayed in heatmap}

\item{plateform}{palteform type (IlluminaHumanMethylation450k, IlluminaHumanMethylationEPIC, MM280)}

\item{genome}{genome assembly}
}
\value{
None
}
\description{
Create a differential methylation analysis report in Rmarkdown.
}
