% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylkey.diff.r
\name{methyldiff}
\alias{methyldiff}
\title{Methylkey differential analysis}
\usage{
methyldiff(
  se = NULL,
  model = NULL,
  intercept = NULL,
  method = "ls",
  niter = 50,
  ncore = 2,
  fdr = 0.05,
  pcutoff = 0.2,
  maxgap = 1000,
  genome = "hg38"
)
}
\arguments{
\item{se}{summarizedExperiment object with betas values in assays and illumina manifest as Features}

\item{model}{model to use for the analysis (string) eg: "~group+gender"}

\item{method}{method to use in regression (ls or robust)}

\item{niter}{number of iteration with robust method}

\item{ncore}{number of core use for regression analysis}

\item{pcutoff}{qvalue cutoff for selecting dmps used for dmr analysis}

\item{modelSVA}{model to use for the batch correction (string) if different from model.}

\item{case}{name of case level}

\item{control}{name of control level}

\item{sva}{if TRUE, sva batch correction will be applied to mvalues}

\item{qval}{qvalue cutoff for selecting significant dmps}

\item{sex}{if FALSE remove XY probes for analysis}
}
\value{
MethylResultSet
}
\description{
Create a differential methylation analysis report in Rmarkdown.
}
