% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\docType{class}
\name{MethylResultSet}
\alias{MethylResultSet}
\alias{MethylResultSet-class}
\title{MethylResultSet Class}
\usage{
MethylResultSet(se, dmps, dmrs = NULL, pca = NULL)

MethylResultSet(se, dmps, dmrs = NULL, pca = NULL)
}
\arguments{
\item{se}{A MethylSet object containing DNA methylation data.}

\item{dmps}{A data frame containing differential methylation positions (DMPs).}

\item{dmrs}{A list of data frames containing differential methylation regions (DMRs).}

\item{pca}{A list containing principal component analysis (PCA) results.}
}
\value{
An object of class MethylResultSet.
}
\description{
A class for storing and working with results from DNA methylation analysis.
The class contains the following slots:
\describe{
\item{dmps}{A data frame (DataFrame) containing differential methylation positions (DMPs).}
\item{dmrs}{A list (List_OR_NULL) containing differential methylation regions (DMRs).}
\item{pca}{A list containing principal component analysis (PCA) results.}
\item{metadata}{A list containing metadata or additional information about the results.}
}

Create an instance of the MethylResultSet class to store DNA methylation analysis results.
}
\section{Slots}{

\describe{
\item{\code{dmps}}{A data frame (DataFrame) containing differential methylation positions (DMPs).}

\item{\code{dmrs}}{A list (List_OR_NULL) containing differential methylation regions (DMRs).}

\item{\code{pca}}{A list containing principal component analysis (PCA) results.}

\item{\code{metadata}}{A list containing metadata or additional information about the results.}
}}

\seealso{
\code{\link{MethylResultSet-class}}
}
\author{
Your Name
}
