% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylKey.r
\name{nbsamples}
\alias{nbsamples}
\title{Generic function to retrieve the number of samples from an object}
\usage{
nbsamples(x)
}
\arguments{
\item{x}{An object from which to retrieve the number of samples.}
}
\value{
The number of samples.
}
\description{
This is a generic function to retrieve the number of samples from an object.
}
