% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_Channels}
\alias{plot_Channels}
\title{Generic function for plotting channels}
\usage{
plot_Channels(x)
}
\arguments{
\item{x}{Input data. Depending on its class, behavior varies.}
}
\value{
A plot representing the channels.
}
\description{
This generic function provides a common interface for plotting channels. The
behavior of this function depends on the class of the input data.
}
