% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_Channels2,RGChannelSet-method}
\alias{plot_Channels2,RGChannelSet-method}
\title{Call Plot channels from an RGChannelSet object for each sentrix}
\usage{
\S4method{plot_Channels2}{RGChannelSet}(x)
}
\arguments{
\item{x}{An RGChannelSet object containing microarray channel data.}
}
\value{
A list of boxplots.
}
\description{
This method of \code{plot_Channels} is specifically for plotting channels
when the input data is an RGChannelSet object. It creates separate boxplots
for each sentrix
}
