% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_Detection}
\alias{plot_Detection}
\title{Create a bar plot to visualize detection fractions}
\usage{
plot_Detection(df)
}
\arguments{
\item{df}{A data frame containing the data to be plotted, with columns 'name' and 'frac_dt'.}
}
\value{
A ggplot2 bar plot for visualizing detection fractions.
}
\description{
This function generates a bar plot to visualize detection fractions using ggplot2. It plots data from a data frame, where 'name' is used on the x-axis and 'frac_dt' on the y-axis.
}
