% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{plot_PCA_contribution}
\alias{plot_PCA_contribution}
\title{Plot PCA Contribution}
\usage{
plot_PCA_contribution(pca, nPC = 4, title = "")
}
\arguments{
\item{pca}{An object of class 'prcomp' containing PCA results.}

\item{nPC}{Number of principal components to include in the plot.}

\item{title}{Title for the plot.}
}
\value{
A ggplot object representing the PCA contribution plot.
}
\description{
Visualizes the contribution of principal components (PC) in a PCA analysis.
}
