% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.r
\name{sesame2Betas}
\alias{sesame2Betas}
\title{Convert Illumina IDAT Files to Beta Values Using sesame}
\usage{
sesame2Betas(
  idat = NULL,
  prep = "QCDPB",
  sampleSheet = NULL,
  na = 0.2,
  ncore = 4
)
}
\arguments{
\item{idat}{The path to the directory containing Illumina IDAT files.}

\item{prep}{The sesame preprocessing method to use. Default is "QCDPB".}

\item{sampleSheet}{A data frame containing sample information, including barcode, sex, and age.}

\item{na}{The maximum proportion of missing values allowed in beta values. Must be between 0 and 1. Default is 0.2.}

\item{ncore}{The number of CPU cores to use for parallel processing. Default is 4.}
}
\value{
An object of class "MethyLumiMethyLumiSet" containing beta values and associated metadata, including quality control statistics and plots.
}
\description{
This function takes Illumina IDAT files, processes them using the sesame package, and returns beta values (DNA methylation values) along with quality control statistics and plots.
}
