% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylResultSet.r
\docType{class}
\name{List_OR_NULL-class}
\alias{List_OR_NULL-class}
\title{Define a Class Union}
\description{
This function allows you to define a class union, which is a special type of class
that includes objects from multiple classes. A class union is useful when you want
to create a class that can accept objects of different classes as its instances.
}
\keyword{classes}
